# data/contracts/openapi.v2.json
{
  "info": {
    "title": "Toonflow Provider API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/files/upload": {
      "post": {
        "operationId": "uploadProviderFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "additionalProperties": false,
                "properties": {
                  "canonicalModelId": {
                    "pattern": "^[a-z0-9]+(?:[._:-][a-z0-9]+)*$",
                    "type": "string"
                  },
                  "idempotencyKey": {
                    "maxLength": 200,
                    "minLength": 8,
                    "type": "string"
                  },
                  "input": {
                    "anyOf": [
                      {
                        "additionalProperties": false,
                        "properties": {
                          "byteLength": {
                            "exclusiveMinimum": 0,
                            "maximum": 67108864,
                            "type": "integer"
                          },
                          "dataBase64": {
                            "minLength": 1,
                            "type": "string"
                          },
                          "expiresAfterSeconds": {
                            "maximum": 2592000,
                            "minimum": 3600,
                            "type": "integer"
                          },
                          "filename": {
                            "maxLength": 255,
                            "minLength": 1,
                            "type": "string"
                          },
                          "mediaType": {
                            "pattern": "^[a-z0-9][a-z0-9!#$&^_.+-]*\\/[a-z0-9][a-z0-9!#$&^_.+-]*$",
                            "type": "string"
                          }
                        },
                        "required": [
                          "dataBase64",
                          "byteLength",
                          "mediaType",
                          "filename"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "assetId": {
                            "pattern": "^sha256:[a-f0-9]{64}$",
                            "type": "string"
                          },
                          "expiresAfterSeconds": {
                            "maximum": 2592000,
                            "minimum": 3600,
                            "type": "integer"
                          },
                          "filename": {
                            "maxLength": 255,
                            "minLength": 1,
                            "type": "string"
                          },
                          "source": {
                            "const": "owned_asset",
                            "type": "string"
                          }
                        },
                        "required": [
                          "source",
                          "assetId"
                        ],
                        "type": "object"
                      }
                    ]
                  },
                  "offeringId": {
                    "pattern": "^[a-z0-9]+(?:[._:-][a-z0-9]+)*$",
                    "type": "string"
                  },
                  "schemaVersion": {
                    "const": "1.0.0",
                    "type": "string"
                  }
                },
                "required": [
                  "schemaVersion",
                  "idempotencyKey",
                  "canonicalModelId",
                  "offeringId",
                  "input"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "byteLength": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "expiresAt": {
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                          "type": "string"
                        },
                        "fileId": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "filename": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "mediaType": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "providerId": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "schemaVersion": {
                          "const": "1.0.0",
                          "type": "string"
                        }
                      },
                      "required": [
                        "schemaVersion",
                        "providerId",
                        "fileId",
                        "mediaType"
                      ],
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Provider-scoped file reference"
          }
        }
      }
    },
...
    "/api/v2/media-assets/upload": {
      "put": {
        "operationId": "uploadOwnedMediaAsset",
        "parameters": [
          {
            "in": "header",
            "name": "X-Toonflow-Media-Type",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Toonflow-Filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "assetId": {
                          "pattern": "^sha256:[a-f0-9]{64}$",
                          "type": "string"
                        },
                        "byteLength": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "filename": {
                          "maxLength": 255,
                          "minLength": 1,
                          "type": "string"
                        },
                        "kind": {
                          "enum": [
                            "image",
                            "video",
                            "audio",
                            "file"
                          ],
                          "type": "string"
                        },
                        "mediaType": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "sha256": {
                          "pattern": "^[a-f0-9]{64}$",
                          "type": "string"
                        }
                      },
                      "required": [
                        "assetId",
                        "kind",
                        "mediaType",
                        "byteLength",
                        "sha256",
                        "filename"
                      ],
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "data",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Principal-owned content-addressed asset"
          }
        }
      }
    },
...
  }
}